requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
motor==3.6.0
pymongo==4.10.1

# Authentication
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# LLM & RAG
langchain==0.3.14
langchain-openai==0.3.0
langchain-anthropic==0.3.6
langchain-google-genai==2.0.8
langchain-mongodb==0.3.0
openai==1.58.1
anthropic==0.42.0
google-genai==1.5.0

# Embeddings & Vector
tiktoken==0.8.0
cohere==5.13.3
sentence-transformers==3.3.1

# Configuration
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Security
python-jose[cryptography]==3.3.0

# Rate limiting
slowapi==0.1.9

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# CORS
starlette==0.41.3
